Partial commit: nicer progress windows, and refactored the 'out string error' pattern into an Exception pattern. Will continue this refactoring.
git-svn-id: http://svn.code.sf.net/p/megui/code@129 a21b8521-4df1-4ed3-b48c-cff5bc73dcb3

+++ trunk/packages/tools/autoencode/AutoEncodeWindow.cs
@@ -853,7 +853,7 @@ namespace MeGUI
             bool cont = info.JobUtil.getFinalZoneConfiguration(vSettings, info.Video.Info.IntroEndFrame, info.Video.Info.CreditsStartFrame);
             if (cont)
             {
-                int length = 0;
+                ulong length = 0;
                 double framerate = 0.0;
                 VideoStream myVideo = new VideoStream();
                 info.JobUtil.getInputProperties(out length, out framerate, info.Video.VideoInput);
